fix: update Number type format to 'double' and enhance SchemaType definition for additional properties

+++ src/transformer.fixtures.ts
@@ -13,7 +13,7 @@ const TS_CONFIG_DEFAULT_PATH = path.resolve(process.cwd(), 'tsconfig.json')
 
 const jsPrimitives = {
   String: { type: 'String', value: 'string' },
-  Number: { type: 'Number', value: 'number', format: 'float' },
+  Number: { type: 'Number', value: 'number', format: 'double' },
   Boolean: { type: 'Boolean', value: 'boolean' },
   Symbol: { type: 'Symbol', value: 'string' },
   BigInt: { type: 'BigInt', value: 'integer', format: 'int64' },
